Refactor SettlementClient to use pydantic types
